Move timeout related function to own module
